core(imports): refactor into imports package and split into modules
