+++ rdmo/core/imports/utils.py
@@ -1,0 +1,26 @@
+import tempfile
+import time
+from os.path import join as pj
+from random import randint
+
+
+def handle_uploaded_file(filedata):
+    tempfilename = generate_tempfile_name()
+    with open(tempfilename, 'wb+') as destination:
+        for chunk in filedata.chunks():
+            destination.write(chunk)
+    return tempfilename
+
+
+def handle_fetched_file(filedata):
+    tempfilename = generate_tempfile_name()
+    with open(tempfilename, 'wb+') as destination:
+        destination.write(filedata)
+    return tempfilename
+
+
+def generate_tempfile_name():
+    t = int(round(time.time() * 1000))
+    r = randint(10000, 99999)
+    fn = pj(tempfile.gettempdir(), 'upload_' + str(t) + '_' + str(r) + '.xml')
+    return fn
